♻ Support different status revision (#118)
* 🐛 Set maxsplit to 4 for status_revision
This is so that if any dashes are used in the status_revision description, they are not split

* 🦺 Relax status_revision validation to be a string
In its current setup, the API will reject any new values that are not in the default enumeration.
By setting to string, we can bypass this and the validation will be done on the frontend.

* 👷 Bump setup-pixi to v0.8.3

* 🔧 Update pixi.lock

* 👷 Bump pixi-version in setup-pixi

* 🦺 Add relaxed validation to check maximum number of dashes